Scrolling Background ver.1
The background has been added and it will scroll infinitely. The character rotates up due to the camera being bound to the character and the background bound to the camera. There is currently no proper floor or walls. The next BG push will have a side wall and camera limits.