Slider is no longer interactive.
Not sure what this did but i can fix it later if it broke anything and I wanted an excuse to push because I changed one menu location and it looks better now. I also forgot to mention I added a box collider to add stage limits. also easily removable if we don't want limits.